move createContextOnPlay to logic
